Add hermetic clang-tidy static analysis for the C/C++ clients
Add a clang_tidy_test rule that runs the LLVM distribution's clang-tidy as
a build action, deriving include/define flags from the target's CcInfo
compilation context (so it resolves the same headers the compiler does),
and ties the action to a test via a stamp. Registered in the client/cpp
and client/cnano lint suites, replacing the cppcheck --check-config no-op.

A checked-in .clang-tidy enables bugprone-* and performance-* over the
hand-written client sources (generated protobuf/service code excluded),
taking over cpplint's non-formatting checks. A handful of checks are
disabled for intentional low-level idioms (documented in the config).

Fixes the real findings this surfaced: for-range-copy and value copies now
take const references (client.cpp, stream_manager.cpp); explicit narrowing
casts in platform.cpp and cnano utils.c/communication.c; the deliberately
empty EncodingError catches (retry-until-enough-bytes) and the no-OOM
recalloc carry documented NOLINTs.

+++ client/cnano/src/utils.c
@@ -7,7 +7,7 @@ void krpc_float64_to_float32(void * float64, void * float32) {
   uint64_t sign = (data >> 32) & 0x80000000;
   uint64_t exp = (data >> 52) & 0x7ff;
   uint64_t man = (data >> 29) & 0x7fffff;
-  int32_t expValue = exp - 1023;
+  int32_t expValue = (int32_t)(exp - 1023);
   if (expValue == 1024) {
     // preserve inf and nan
     expValue = 128;
@@ -28,7 +28,7 @@ void krpc_float32_to_float64(void * float32, void * float64) {
   uint64_t sign = ((data >> 31) & 0x1) << 63;
   uint64_t exp = (data >> 23) & 0xff;
   uint64_t man = (data & 0x7fffff) << 29;
-  int32_t expValue = exp - 127;
+  int32_t expValue = (int32_t)(exp - 127);
   if (expValue == -127)
     expValue = -1023;
   if (expValue == 128)
